+++ src/components/Box/index.js
@@ -1,5 +1,5 @@
 import { useState, useRef } from 'react'
-import useOnClickOutside from '../../../hooks/useOnClickOutside'
+import useOnClickOutside from '../../hooks/useOnClickOutside'
 
 import BtnAddCategory from '../BtnAddCategory'
 import AddCategoryForm from '../AddCategoryForm'
