+++ src/components/BtnClose/BtnClose.module.scss
