+++ src/components/ColorList/index.js
