+++ src/components/SelectedCategoryForm/SelectedCategoryForm.module.scss
